fix: Add missing exports due to merge conflict.
